requirements.txt
Django==2.2
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
Pillow==9.0.0
pytz==2021.3
sqlparse==0.4.2
